utils/Yume/pyproject.toml
[project]
name = "yume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "jinja2>=3.1.6,<4",
    "markdown>=3.10,<4",
    "pillow>=12.1.0,<13",
]
